добавляет отдельный модуль эффектов
